+++ lib/sudo-pam/Cargo.toml
@@ -10,5 +10,6 @@ publish.workspace = true
 [dependencies]
 sudo-pam-sys.workspace = true
 sudo-cutils.workspace = true
+sudo-system.workspace = true
 thiserror.workspace = true
 libc.workspace = true
