+++ lib/sudo-pam/src/lib.rs
@@ -6,14 +6,13 @@ use std::{
 use converse::{Converser, ConverserData};
 use error::pam_err;
 pub use error::{PamError, PamErrorType, PamResult};
-use sudo_cutils::{string_from_ptr, Secure};
+use sudo_cutils::string_from_ptr;
 use sudo_pam_sys::*;
 
 mod converse;
 mod error;
 mod rpassword;
-
-pub(crate) type Password = Secure<Vec<u8>>;
+mod securemem;
 
 pub use converse::CLIConverser;
